mobile_getAllConnections
