+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/UserCommander.scala
@@ -11,7 +11,7 @@ import play.api.libs.json._
 import com.google.inject.Inject
 import com.keepit.common.social.BasicUserRepo
 import com.keepit.social.{UserIdentity, SocialNetworks, BasicUser}
-import scala.concurrent.Future
+import scala.concurrent.{Await, Future}
 import com.keepit.common.usersegment.UserSegment
 import com.keepit.common.usersegment.UserSegmentFactory
 import scala.util.{Failure, Success, Try}
@@ -20,6 +20,8 @@ import com.keepit.common.akka.SafeFuture
 import com.keepit.heimdal.{UserEventTypes, UserEvent, HeimdalServiceClient, HeimdalContextBuilderFactory}
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import securesocial.core.{Identity, UserService, Registry}
+import java.text.Normalizer
+import com.keepit.common.logging.Logging
 
 
 case class BasicSocialUser(network: String, profileUrl: Option[String], pictureUrl: Option[String])
@@ -70,9 +72,11 @@ class UserCommander @Inject() (
   bookmarkRepo: BookmarkRepo,
   userExperimentRepo: UserExperimentRepo,
   socialUserInfoRepo: SocialUserInfoRepo,
+  socialConnectionRepo: SocialConnectionRepo,
+  invitationRepo: InvitationRepo,
   eventContextBuilder: HeimdalContextBuilderFactory,
   heimdalServiceClient: HeimdalServiceClient,
-  abook: ABookServiceClient) {
+  abookServiceClient: ABookServiceClient) extends Logging {
 
   def getFriends(user: User, experiments: Set[ExperimentType]): Set[BasicUser] = {
     val basicUsers = db.readOnly { implicit s =>
@@ -113,7 +117,7 @@ class UserCommander @Inject() (
   }
 
   def uploadContactsProxy(userId: Id[User], origin: ABookOriginType, payload: JsValue): Future[JsValue] = {
-    abook.uploadContacts(userId, origin, payload)
+    abookServiceClient.uploadContacts(userId, origin, payload)
   }
 
   def getUserInfo(user: User): BasicUserInfo = {
@@ -181,6 +185,91 @@ class UserCommander @Inject() (
       }
     }
     resOpt getOrElse { throw new IllegalArgumentException("no_user") }
+  }
+
+  @inline def normalize(str: String) = Normalizer.normalize(str, Normalizer.Form.NFD).replaceAll("\\p{InCombiningDiacriticalMarks}+", "").toLowerCase
+
+  private def queryContacts(userId:Id[User], search: Option[String], after:Option[String], limit: Int):Future[Seq[JsObject]] = { // TODO: optimize
+
+    @inline def mkId(email:String) = s"email/$email"
+    @inline def getEInviteStatus(contactIdOpt:Option[Id[EContact]]):String = { // todo: batch
+      contactIdOpt flatMap { contactId =>
+        db.readOnly { implicit s =>
+          invitationRepo.getBySenderIdAndRecipientEContactId(userId, contactId) map { inv =>
+            if (inv.state != InvitationStates.INACTIVE) "invited" else ""
+          }
+        }
+      } getOrElse ""
+    }
+
+    abookServiceClient.queryEContacts(userId, limit, search, after) map { paged =>
+      val objs = paged.take(limit).map { e =>
+        Json.obj("label" -> JsString(e.name.getOrElse("")), "value" -> mkId(e.email), "status" -> getEInviteStatus(e.id))
+      }
+      log.info(s"[queryContacts(id=$userId)] res(len=${objs.length}):${objs.mkString.take(200)}")
+      objs
+    }
+  }
+
+  def getAllConnections(userId:Id[User], search: Option[String], network: Option[String], after: Option[String], limit: Int):Future[Seq[JsObject]] = { // todo: convert to objects
+    val contactsF = if (network.isDefined && network.get == "email") { // todo: revisit
+      queryContacts(userId, search, after, limit)
+    } else Future.successful(Seq.empty[JsObject])
+    @inline def socialIdString(sci: SocialConnectionInfo) = s"${sci.networkType}/${sci.socialId.id}"
+    val searchTerms = search.toSeq.map(_.split("\\s+")).flatten.filterNot(_.isEmpty).map(normalize)
+    @inline def searchScore(sci: SocialConnectionInfo): Int = {
+      if (network.exists(sci.networkType.name !=)) 0
+      else if (searchTerms.isEmpty) 1
+      else {
+        val name = normalize(sci.fullName)
+        if (searchTerms.exists(!name.contains(_))) 0
+        else {
+          val names = name.split("\\s+").filterNot(_.isEmpty)
+          names.count(n => searchTerms.exists(n.startsWith))*2 +
+            names.count(n => searchTerms.exists(n.contains)) +
+            (if (searchTerms.exists(name.startsWith)) 1 else 0)
+        }
+      }
+    }
+
+    def getWithInviteStatus(sci: SocialConnectionInfo)(implicit s: RSession): (SocialConnectionInfo, String) =
+      sci -> sci.userId.map(_ => "joined").getOrElse {
+        invitationRepo.getByRecipientSocialUserId(sci.id) collect {
+          case inv if inv.state != InvitationStates.INACTIVE => "invited"
+        } getOrElse ""
+      }
+
+    def getFilteredConnections(sui: SocialUserInfo)(implicit s: RSession): Seq[SocialConnectionInfo] =
+      if (sui.networkType == SocialNetworks.FORTYTWO) Nil
+      else socialConnectionRepo.getSocialConnectionInfo(sui.id.get) filter (searchScore(_) > 0)
+
+    val connections = db.readOnly { implicit s =>
+      val filteredConnections = socialUserInfoRepo.getByUser(userId)
+        .flatMap(getFilteredConnections)
+        .sortBy { case sui => (-searchScore(sui), normalize(sui.fullName)) }
+
+      (after match {
+        case Some(id) => filteredConnections.dropWhile(socialIdString(_) != id) match {
+          case hd +: tl => tl
+          case tl => tl
+        }
+        case None => filteredConnections
+      }).take(limit).map(getWithInviteStatus)
+    }
+
+    val jsConns: Seq[JsObject] = connections.map { conn =>
+      Json.obj(
+        "label" -> conn._1.fullName,
+        "image" -> Json.toJson(conn._1.getPictureUrl(75, 75)),
+        "value" -> socialIdString(conn._1),
+        "status" -> conn._2
+      )
+    }
+    contactsF map { jsContacts =>
+      val jsCombined = jsConns ++ jsContacts
+      log.info(s"[getAllConnections(${userId})] jsContacts(sz=${jsContacts.size}) jsConns(sz=${jsConns.size})")
+      jsCombined
+    }
   }
 
 }
